Move components into component folders
