+++ src/Components/Resources/content.js
