+++ src/Components/Rights/Rights.jsx
