+++ src/Components/Translate/translate.scss
