+++ src/scss/variables.scss
@@ -22,3 +22,5 @@ $border-radius-lg: 1rem;
 $border-radius-xl: 1rem;
 $border-radius-xxl: 1rem;
 $border-radius-2xl: 1rem;
+
+$section-margin-bottom: 3rem;
